sched: qhmp: add a knob to restrict tasks spreading
There is an opportunity to save power by consolidating the load
to fewer CPUs on some targets. The results indicate that the
aggressive packing results in higher frequency residency, but
still help power as fewer CPUs are used. Provide a knob to
restrict the task spreading.

When sched_restrict_tasks_spread knob is enabled tasks spreading
is contained by applying the following two policies.

- The small tasks are packed up to spill thresholds which otherwise
are packed up to mostly_idle thresholds. This helps in consolidating
the load to fewer CPUs.

- The current CPU selection algorithm for RT tasks looks for the
least loaded CPU in the lower power cluster. Due to this, RT tasks
are getting placed on the idle CPU more often. This CPUs pulls the
tasks from other CPUs as part of the newly_idle load balance after
finishing the RT task execution. This results in more concurrent
CPU usage and hurting the power. Restrict the search to first
available lower power CPU to contain the load to fewer CPUs.
When sched_boost is ON, this restriction is not applied.

CRs-Fixed: 996652
Change-Id: If0d481cc74ff084fe72ff6cdc4d6aeba96171653
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: RenderBroken <zkennedy87@gmail.com>